requirements.txt
Django==3.1.4
django-pygmentify==0.3.7
django-ckeditor==6.0.0
django-blacklist==0.4.1
django-cleanup-5.2.0==5.2.0
Pillow==8.2.0
djangorestframework==3.12.4
django-ratelimit==3.0.1
